supress warning.
git-svn-id: https://cutter.svn.sourceforge.net/svnroot/cutter/trunk@1651 10a583ed-1e3d-0410-b285-c80daaaae891

+++ cutter/cut-sequence-matcher.c
@@ -257,7 +257,7 @@ char_sequence_new (const gchar *string)
     sequence = g_sequence_new(NULL);
 
     for (; *string != '\0'; string++) {
-        g_sequence_append(sequence, GINT_TO_POINTER(*string));
+        g_sequence_append(sequence, GINT_TO_POINTER((gint)*string));
     }
 
     return sequence;
